-am rezolvat problema cu keya creata de hibernate mapasem gresit in baza de date o variabila -place order functional pe back -modificare entitati pt tipul de user -adaugare entitate in persistance pt hibernate

+++ src/main/java/dto/ProductDTO.java
@@ -6,7 +6,7 @@ public class ProductDTO {
 
     private String productName;
 
-    private int stockNumber;
+    private Long stockNumber;
 
     private String category;
 
@@ -29,11 +29,11 @@ public class ProductDTO {
         this.productName = productName;
     }
 
-    public int getStockNumber() {
+    public Long getStockNumber() {
         return stockNumber;
     }
 
-    public void setStockNumber(int stockNumber) {
+    public void setStockNumber(Long stockNumber) {
         this.stockNumber = stockNumber;
     }
